==> Models/4 train test split VM.ipynb <==
import pandas as pd
import numpy as np
from tqdm.notebook import tqdm

action_num = 16
version = '240721'
action_dim = action_num

data = pd.read_csv(f'./Shock dataset 1hr state_cont_{version}_Action{action_num}_Reward Mean GAP VM.csv')

data

import random

# Random seed 고정
random.seed(220604)
np.random.seed(220604)

traj = data['stay_id'].unique().tolist()
num_80 = len(traj) * 0.8
int_num_80 = round(num_80)
random.shuffle(traj)
train_traj = traj[:int_num_80]
test_traj = traj[int_num_80:]

train_data = pd.DataFrame()
val_data = pd.DataFrame()

for p in train_traj:
    selected1 = data[data['stay_id'] == p].reset_index(drop=True)
    train_data = pd.concat([train_data, selected1], ignore_index=True)

for q in test_traj:
    selected1 = data[data['stay_id'] == q].reset_index(drop=True)
    val_data = pd.concat([val_data, selected1], ignore_index=True)

print(len(train_data[train_data['Death'] == 1]))
print(len(val_data[val_data['Death'] == 1]))

train_data.to_csv(f'./train_shock_dataset_state_cont_{version}_Action{action_num}_Reward_Mean_VM.csv', index=False)
val_data.to_csv(f'./val_shock_dataset_state_cont_{version}_Action{action_num}_Reward_Mean_VM.csv', index=False)
